# Avoid making a subscope in tallyMetricsHandler if possible