Creating a subscope, even if it is `tmp.Scope.Tagged(nil)`, takes extra time.

Used this script to [benchmark](https://gist.github.com/k24dizzle/f652990a342214c539d348c46e831f41).

Before:
```
go test -benchmem -run=^$ -bench . -benchtime=10s
goos: darwin
goarch: arm64
pkg: test/prom13
BenchmarkCounter-10     67209068               170.4 ns/op            48 B/op          2 allocs/op
PASS
ok      test/prom13     11.812s
```

After:
```
go test -benchmem -run=^$ -bench . -benchtime=10s
goos: darwin
goarch: arm64
pkg: test/prom13
BenchmarkCounter-10     84332931               138.2 ns/op            48 B/op          2 allocs/op
PASS
ok      test/prom13     12.028s
```